Python program to calculate income tax on annual income.
Write a code to calculate income tax on annual income as follows. 
For first 250000 income tax is 0%, 
For Next 250000 income tax is 5%, 
For Next 500000 income tax is 10% and 
For remaining amount income tax is 30%. 
Calculate and display final tax payable. (Code should work for tax calculation on any amount ) 
Upload .py file of your code here and turn-in/hand-in the assignment. 
(To get .py file of your code in jupyter notebook, goto file menu-> Download as-> Python)


Test Cases:
               INPUT      OUTPUT
• 180000         0
• 400000         7500
• 800000         42500
• 1200000       122500