Write a python code to accept marks from user and display final marks.
In a college grading system, marks are given in 0-100 range. As a final grade, students marks should be rounded to multiple of 5. Rounding must be done using following rules. 
A. If the difference between the current marks and the next multiple of 5 is less than 3, round marks up to the next multiple of 5 else round down to previous multiple of 5. 
B. Do not round if Student is failed (<40) 
Examples 
73 =>round to 75 (75 - 73 is less than 3,, so round to next multiple of 5=>75) 
30 =>do not round (result is less than 40) 
52 =>round to 50 (55 - 52 is not less than 3, so round to previous multiple of 5=>50) 
. Upload .py file of your code here and Turn in the assignment.


TEST CASES:


       INPUT     OUTPUT
56        55
58        60
39        39
92        90