Password Validation
A website requires the users to input username and password to register. Write a program to check the validity of password input by users. Following are the criteria for checking the password:
1. At least 1 letter between [a-z]
2. At least 1 number between [0-9]
3. At least 1 letter between [A-Z]
4. At least 1 character from [$,#,@]
5. Minimum length of password: 6
6. Maximum length of password: 12
Your program should accept a password from user and will check it according to the above criteria. Password that match the criteria are to be printed as valid else print invalid.
Example If the following passwords are given as input to the program:
ABd1234@1 => VALID
F1#2w3E* => VALID
2We3345 => INVALID